Add condition for destroy container.

+++ destroyOpenLDAP.sh
@@ -2,5 +2,7 @@
 
 LDAP_NAME=${LDAP_NAME:-openldap}
 
+if [ -n "$(docker ps -a | grep ${LDAP_NAME})" ]; then
 docker stop ${LDAP_NAME}
 docker rm -v ${LDAP_NAME}
+fi
